solution for this problem is solved in the pdf
 A dynamic vibration absorber is shown in Figure 1. This system 
is representative of many situations involving the vibration of machines containing 
unbalanced components. The parameters M2 and k12 may be chosen so that the main 
mass M1 does not vibrate in the steady-state when F(t) = 2 * sin (10 * t)
Obtain the differential equations describing this system.
Simulate this system for 10 seconds. M1 = 100. K1 = 50. b = 50.
Find the optimal value for M2 and k12 so that M1 does not vibrate.